constructor works, green again
we wont expose attrs, but we will have methods that inform the robot if
its facing an edge. this will allow us to write code like

robot.is_facing_edge?

for when the robot trying to make decisions about whether to move or
not.
safe robots (as required by the challenge) will not move if that is true
unsafe robots will move

we'll model these next