follow Sandi Metz's guide on testing command methods and collaborators
she recommends, for commands that work with collaborators, to confirm
that the collaborator is called, via a mock (calls that leave the space
capsule, if you've seen the video).

we do that here. we could use a verifying double, to make sure we catch
any drift in the collaborators API, and we've updated the spec_helper
for that, but we wont get too fancy right now.