conveying an understanding of Ruby's dynamic nature.
i'm reading the PDF spec and taking as a constraint of the Robot class
(referencing the class-contract ideas present in the Eiffel language),
that all commands before the robot is placed to be ignored.

now we could model state with a @placed attribute, and the check this
before executing all the other commands.

or we could have the commands not even exist until we have successfully
placed the robot.

those 'def xxx' statements _inside_ the .place method do not get
executed until the table has agreed we are placing the robot in a safe
place.

it can be quite strongly argued that having parts of an objects API
spring into existence as the result of a method call is unusual, but in
this case its a nice illustration of enforcing an constraint without
sprinkling "if @placed" controls or "has_been_placed? {command_impl}"
callbacks.

i've committed without tests so you can see the basics. its the first
time i've used this mechanism, i'm interested to see how the tests play
out.