robots get the other command methods once they have been placed.
the methods have to become singleton methods on the specific robot
instance, otherwise they exist as instance methods of the class, and any
subsequent unplaced robots (such as the new ones created for each test)
get the command methods as soon as one robot gets them.

could have used the 'class << self' style as well, as in this context
self is the robot instance.