method_missing returns exceptions, not strings
let the user determine the policy of how to handle these cases.

this allows the user of a robot to decide how to handle the failures,
rather than hard-wiring in just throwing them all away. the user could
catch the UnplacedErrors and keep reading commands, or they could exit
immediately, or they could re-throw.

we could fiddle with the backtrace to make the exceptions appear to come
from the callers code, but not for this exercise