tdd red - what to do ?
so we're at the point where we can build out the facing attribute of a
robot, but to do so we change the signature of the public #place()
method.

this breaks a bunch of tests, but before we go on a big refactor we make
note that .place() had always needed a facing param, but we had ignored
until we were ready. So its not appropriate to invoke a discussion on
how we have violated the 'O' in SOLID, as we haven't reached the point
where we can consider the code closed for modification.

we'll be lazy and fix the tests by defaulting facing to north in the
next commit.

also remove the silly array intersection code for array membership
testing - the Array#include? had slipped my mind last night.