left and right turning tests
we combined the list of facings and the current facing into the same
structure, to demonstrate an understanding of array rotation. so we
cycle the list when turning, so that the current facing is always the
first in the list. so the list does double duty as a validator and a
holder of state. you would almost certainly never do this in 'real' code

we also demonstrate breaking encapsulation by querying a private method
from the tests - just to show we know how.