parser is enumerable - sorta
so i've kinda shot myself in the foot by trying to show off - but it
works for the cases i need it to. i thought Enumerable mixed in more
methods than it does, so the interface isnt as clean as i hoped and
there will be a bit of jiggery-pokery next as we build the command
runner. the enumerator also has the nasty destructive list processing
(the call to shift), but i'm not willing right now to spend the time
removing this unwanted side-effect. i also wont spend the time to
incorporate lazy enumeration and have the robot process potentially
infinite-length command lists. i think mentioning it is enough to show
my thought-processes.